% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkRUnBBayes}
\alias{checkRUnBBayes}
\title{Check if the main functions of runbbyes are running correctly.}
\usage{
checkRUnBBayes()
}
\value{
TRUE if everything runs smoothly, FALSE otherwise.
}
\description{
Check if the main functions of runbbyes are running correctly.
}

