% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hasNode}
\alias{hasNode}
\title{Has node}
\usage{
hasNode(event, name)
}
\arguments{
\item{event}{a list of tuples <node, state>}

\item{name}{the node's name to search for}
}
\value{
TRUE if the name is found in the event list, FALSE otherwise.
}
\description{
Search an existing node
}

