% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.network}
\alias{print.network}
\title{prints the network information in runbbayes.}
\usage{
\method{print}{network}(x, ...)
}
\arguments{
\item{x}{the network information in runbbayes to be printed.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
prints the network information in runbbayes.
}

