% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.nodeinfo}
\alias{print.nodeinfo}
\title{prints the node information in runbbayes.}
\usage{
\method{print}{nodeinfo}(x, ...)
}
\arguments{
\item{x}{the node information in runbbayes to be printed.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
prints the node information in runbbayes.
}

