\name{print.query}
\alias{print.query}
\title{prints the result of a query in runbbayes.}
\usage{
  print.query(x, ...)
}
\arguments{
  \item{x}{the result of a query in runbbayes to be
  printed.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\description{
  prints the result of a query in runbbayes.
}

