\name{parentsDefined}
\alias{parentsDefined}
\title{Parents defined}
\usage{
parentsDefined(netJava, node)
}
\arguments{
  \item{netJava}{an existing network}

  \item{node}{a list with the node's information returned
  by createNodeInfo function}
}
\description{
Verifies if the node's parents exist.
}

