\name{print.network}
\alias{print.network}
\title{prints the network information in runbbayes.}
\usage{
  print.network(x, ...)
}
\arguments{
  \item{x}{the network information in runbbayes to be
  printed.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\description{
  prints the network information in runbbayes.
}

