\name{print.nodeinfo}
\alias{print.nodeinfo}
\title{prints the node information in runbbayes.}
\usage{
  print.nodeinfo(x, ...)
}
\arguments{
  \item{x}{the node information in runbbayes to be
  printed.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\description{
  prints the node information in runbbayes.
}

