% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.query}
\alias{print.query}
\title{prints the result of a query in runbbayes.}
\usage{
\method{print}{query}(x, ...)
}
\arguments{
\item{x}{the result of a query in runbbayes to be printed.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
prints the result of a query in runbbayes.
}

