% Generated by roxygen2 (4.0.1): do not edit by hand
\name{removeNode}
\alias{removeNode}
\title{Remove node}
\usage{
removeNode(network, name)
}
\arguments{
\item{network}{An existing network}

\item{name}{Node's name}
}
\description{
Removes node of an existing network.
}
\examples{
node.a = createNodeInfo(~asia, prob=c(0.01, 0.99), states=c("yes","no"))
node.t = createNodeInfo(~tub|asia, prob=c(0.05, 0.95, 0.01, 0.99),states=c("yes","no"))
node.s = createNodeInfo(~smoke, prob=c(0.5,0.5), states=c("yes","no"))
node.l = createNodeInfo(~lung|smoke, prob=c(0.1, 0.9, 0.01, 0.99), states=c("yes","no"))
node.b = createNodeInfo(~bronc|smoke, prob=c(0.6, 0.4, 0.3, 0.7), states=c("yes","no"))
node.e = createNodeInfo(~either|lung:tub,prob=c(1,0,1,0,1,0,0,1),states=c("yes","no"))
node.x = createNodeInfo(~xray|either, prob=c(0.98, 0.02, 0.05, 0.95), states=c("yes","no"))
node.d = createNodeInfo(~dysp|bronc:either, prob=c(0.9, 0.1, 0.7, 0.3, 0.8, 0.2, 0.1, 0.9), states=c("yes", "no"))

nodes = list(node.a, node.t, node.s, node.l, node.b, node.e, node.x, node.d)
net = createNetwork(nodes, compile = TRUE)

# Adding a node
net = addNode(net, ~asthma|smoke, prob = c(0.6, 0.4, 0.85, 0.15), states = c("yes", "no"))
net

# Removing a node
net = removeNode(net, "asthma")
net
}

